Restructuring done
